ads1220: Expose registers for FIR filter, power switching, etc. as configuration variables.
This provides the registers of the ADS1220 as configurable parameters for the end user.

Different implementations of the ADC will have varying use cases, therefore it is worthwhile to have these available.

Especially useful is the IDAC, when no external voltage reference is present on the `input_mux` line. For example, when the user is using the built-in source, or the analog supply line, they can use the IDAC to set a predetermined amount of current to be used for reading across a bridge.

The ADS1220 also contains a built-in FIR filter for 50/60Hz noise rejection, which can prove very useful if the ADC is situated close to a switching mode power supply. For load cell probing, this is also done in a software filter, as other chips may lack this capability. Specifically for the ADS1220 however, it should speed up the processing time of load cell probe analysis by using the hardware filter.

Signed-off-by: Ella Fox <ella@fox.gal>